Test onnx==1.14.0 (#988)
* Test onnx==1.14.0


* fix yml syntax


* fix requirements


* fix broken unit test


* fix ir_version


* fix misspelling


* fix error


* update to rc3


* change np.float to np.float32


* Test onnx==1.14.0


* fix yml syntax


* fix requirements


* fix broken unit test


* fix ir_version


* fix misspelling


* fix error


* update to rc3


* change np.float to np.float32


* use official release


* disable documentation on one CI


* disable test for sklearn==1.0.2


* lint


* fix issue for an old scikit-learn on windows


* update CI


* pin numpy version for old build on windows


* resolve comment


* update


---------

+++ skl2onnx/operator_converters/label_binariser.py
@@ -30,9 +30,9 @@ def convert_sklearn_label_binariser(scope: Scope, operator: Operator,
                            name=scope.get_unique_operator_name('identity'))
     else:
         zeros_tensor = np.full((1, len(classes)),
-                               binariser_op.neg_label, dtype=np.float)
+                               binariser_op.neg_label, dtype=np.float32)
         unit_tensor = np.full((1, len(classes)),
-                              binariser_op.pos_label, dtype=np.float)
+                              binariser_op.pos_label, dtype=np.float32)
 
         classes_tensor_name = scope.get_unique_variable_name('classes_tensor')
         equal_condition_tensor_name = scope.get_unique_variable_name(
